#99 try C.size_t

+++ g2configmgr/g2configmgr.go
@@ -133,7 +133,7 @@ func (client *G2configmgr) getLastException(ctx context.Context) (string, error)
 		defer func() { client.traceExit(14, result, err, time.Since(entryTime)) }()
 	}
 	stringBuffer := client.getByteArray(initialByteArraySize)
-	C.G2ConfigMgr_getLastException((*C.char)(unsafe.Pointer(&stringBuffer[0])), C.ulong(len(stringBuffer)))
+	C.G2ConfigMgr_getLastException((*C.char)(unsafe.Pointer(&stringBuffer[0])), C.size_t(len(stringBuffer)))
 	// if result == 0 { // "result" is length of exception message.
 	// 	err = client.getLogger().Error(4006, result, time.Since(entryTime))
 	// }
